refactor: move posting handlers to router
